% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/achim.R
\docType{data}
\name{achim}
\alias{achim}
\title{A list of seven correlation matrices.}
\format{
A list of correlation matrices.
}
\source{
\url{https://github.com/quantmeth}
}
\usage{
achim
}
\description{
A list of seven correlation matrices.
}
\references{
Achim, A. (personal communication).
}
\keyword{datasets}
