% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BartlettSphericity.R
\name{BartlettSphericity}
\alias{BartlettSphericity}
\title{Bartlett Sphericity Test}
\usage{
BartlettSphericity(R, n)
}
\arguments{
\item{R}{the correlation matrix.}

\item{n}{the sample size.}
}
\value{
The \eqn{\chi^2} test of the correlation matrix \code{R} with sample size \code{n}.
}
\description{
\code{BartlettSphericity} tests if variables are orthogonal.
}
\examples{
BartlettSphericity(ex_4factors_corr, 42)
}
\references{
Bartlett, M. S. (1937). Properties of sufficiency and statistical tests. \emph{Proceedings of the Royal Statistical Society, Series A}, \emph{160}, 268–282
}
\author{
André Achim (Matlab)

P.-O. Caron (R)
}
