\name{birdgroupings}
\alias{birdgroupings}
\docType{data}
\title{
Bird Characteristics}
\description{
Displays several characteristics of birds breeding in Canada.}
\usage{data(birdgroupings)}
\format{
  A data frame with 436 observations on the following 48 variables.
  \describe{
    \item{\code{Species_ID}}{a character vector}
    \item{\code{English_Name}}{a character vector}
    \item{\code{French_Name}}{a character vector}
    \item{\code{Scientific_Name}}{a character vector}
    \item{\code{MBCA_2011_Final}}{a character vector}
    \item{\code{mclutch}}{a numeric vector}
    \item{\code{cavity}}{a numeric vector}
    \item{\code{Species_gr1}}{a character vector}
    \item{\code{Species_gr2}}{a character vector}
    \item{\code{Habitat1}}{a character vector}
    \item{\code{Habitat2}}{a character vector}
    \item{\code{Habitat3}}{a character vector}
    \item{\code{Habitat4}}{a character vector}
    \item{\code{Nest_Type1}}{a character vector}
    \item{\code{Nest_Type2}}{a character vector}
    \item{\code{Nest_Type3}}{a character vector}
    \item{\code{Mig_W}}{a character vector}
    \item{\code{Mig_C}}{a character vector}
    \item{\code{Mig_E}}{a character vector}
    \item{\code{size}}{a character vector}
    \item{\code{CAV}}{a numeric vector}
    \item{\code{COL}}{a numeric vector}
    \item{\code{PAR}}{a numeric vector}
    \item{\code{Nest_Reused}}{a logical vector}
    \item{\code{BCR03}}{a logical vector}
    \item{\code{BCR04}}{a logical vector}
    \item{\code{BCR05}}{a logical vector}
    \item{\code{BCR06}}{a logical vector}
    \item{\code{BCR07}}{a logical vector}
    \item{\code{BCR08}}{a logical vector}
    \item{\code{BCR09}}{a logical vector}
    \item{\code{BCR10}}{a logical vector}
    \item{\code{BCR11}}{a logical vector}
    \item{\code{BCR12}}{a logical vector}
    \item{\code{BCR14}}{a logical vector}
    \item{\code{TF}}{a numeric vector}
    \item{\code{TO}}{a numeric vector}
    \item{\code{TU}}{a numeric vector}
    \item{\code{WE}}{a numeric vector}
    \item{\code{SS}}{a numeric vector}
    \item{\code{WAT}}{a numeric vector}
    \item{\code{GRO}}{a numeric vector}
    \item{\code{BOT}}{a numeric vector}
    \item{\code{CLI}}{a numeric vector}
    \item{\code{MAN}}{a numeric vector}
    \item{\code{Family_Sci}}{a character vector}
    \item{\code{Family_CmName_EN}}{a character vector}
    \item{\code{AOU_Sort_Order}}{a numeric vector}
  }
}
\details{
Used to build bird groups}
\source{
To be described}
\references{
To be described}

