% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BOOTR.R
\name{BOOTR}
\alias{BOOTR}
\title{BOOTR}
\usage{
BOOTR(
  D,
  alpha = 0.05,
  nreps = 1000,
  max.fact = max.fact <- max(which(dof(nv, 0:nv) > (nv/2))) - 1
)
}
\arguments{
\item{D}{data}

\item{alpha}{Type I error}

\item{nreps}{Number of replications}

\item{max.fact}{max number of factors}
}
\value{
The number of factors
}
\description{
BOOTR
}
\examples{
\dontrun{
BOOT(ex_2factors)
}
}
