% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/briggs_maccallum2003.R
\docType{data}
\name{briggs_maccallum2003}
\alias{briggs_maccallum2003}
\title{A list of three correlation matrices.}
\format{
A a list of three correlation matrices.
}
\source{
\url{https://github.com/quantmeth}
}
\usage{
briggs_maccallum2003
}
\description{
A list of three correlation matrices.
}
\references{
Briggs, N. E., & MacCallum, R. C. (2003). Recovery of weak common factors by Maximum likelihood and ordinary least squares estimation. \emph{Multivariate Nehavioral Research}, \emph{38}(1), 25–56. \doi{10.1207/S15327906MBR3801_2}
}
\keyword{datasets}
