% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/caron2019.R
\docType{data}
\name{caron2019}
\alias{caron2019}
\title{A list of 15 correlation matrices composed of nine variables with three factors and different levels of correlations between factors.}
\format{
A list of 15 9 x 9 correlation matrices.
}
\source{
\url{https://github.com/quantmeth}
}
\usage{
caron2019
}
\description{
A list of 15 correlation matrices composed of nine variables with three factors and different levels of correlations between factors.
}
\references{
Caron, P.-O. (2019). Minimum average partial correlation and parallel analysis : The influence of oblique structures. \emph{Communications in Statistics - Simulation and Computation}, \emph{48}(7), 2110-2117. \doi{10.1080/03610918.2018.1433843}
}
\keyword{datasets}
