% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cormat.R
\docType{data}
\name{cormat}
\alias{cormat}
\title{A list containing 120 correlation matrices.}
\format{
A a list of 120 correlation matrices.
}
\source{
\url{https://github.com/quantmeth}
}
\usage{
cormat
}
\description{
A list containing 120 \eqn{24 \times 24} correlation matrices 
(R) built to represent different factor structures.
Details are found in the `cormat.l` data.
}
\references{
Caron, P.-O. (2025). A comparison of the Next Eigenvalue Sufficiency Test to other stopping rules for the number of factors in factor analysis. \emph{Educational and Psychological Measurement}. \doi{10.1177/00131644241308528}
}
\keyword{datasets}
