% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/covFIML.R
\name{covFIML}
\alias{covFIML}
\alias{corFIML}
\title{Full Information Maximum Likelihood (FIML) correlation or covariance matrix}
\usage{
covFIML(data, tol = 1e-6, maxiter = 1000, pvalue = FALSE)

corFIML(data, tol = 1e-6, maxiter = 1000, pvalue = FALSE)
}
\arguments{
\item{data}{a data frame of rdata matrix.}

\item{tol}{tolerance.}

\item{maxiter}{maximum number of iterations.}

\item{pvalue}{an argument to indicate if \eqn{p}-values are required.}
}
\value{
A list containing the means, th correlation or covariance matrix, and optionnaly the degree of freedom and the p-values.
}
\description{
Full Information Maximum Likelihood (FIML) correlation or covariance matrix
}
\note{
A not so efficient function. See \code{?cor_nest} instead.
}
\examples{
covFIML(airquality)
}
