\name{ecodistrictscsv}
\alias{ecodistrictscsv}
\docType{data}
\title{
Ecodistrict database
}
\description{
Ecodistricts database
}
\usage{data(ecodistrictscsv)}
\format{
  A data frame with 711218 observations on the following 7 variables.
  \describe{
    \item{\code{long}}{a numeric vector}
    \item{\code{lat}}{a numeric vector}
    \item{\code{order}}{a numeric vector}
    \item{\code{hole}}{a logical vector}
    \item{\code{piece}}{a numeric vector}
    \item{\code{group}}{a numeric vector}
    \item{\code{id}}{a numeric vector}
  }
}
\details{
...
}
\source{
...
}
\references{
...
}
\examples{
data(ecodistrictscsv)
}
\keyword{datasets}
