% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ex.mqr.R
\docType{data}
\name{ex.mqr}
\alias{ex.mqr}
\title{A correlation matrix from chapter 19 Explorer of Méthodes quantitatives avec R (MQR).}
\format{
A 6 by 6 correlation matrix
}
\source{
\url{https://github.com/quantmeth}
}
\usage{
ex.mqr
}
\description{
A population correlation matrix composed of 6 items from a two factor stucture. Factor 1 is based on items 1 to 4, and Factor 2 is based on items 4 to 6.
}
\keyword{datasets}
