% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ex_2factors.R
\docType{data}
\name{ex_2factors}
\alias{ex_2factors}
\title{A correlation matrix composed of 2 factors.}
\format{
A 10 by 10 correlation matrix
}
\source{
\url{https://github.com/quantmeth}
}
\usage{
ex_2factors
}
\description{
A correlation matrix composed of 10 items based on 2 factors with 5 variables
each and loadings equals to .80.
}
\keyword{datasets}
