% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ex_3factors_doub_unique.R
\docType{data}
\name{ex_3factors_doub_unique}
\alias{ex_3factors_doub_unique}
\title{A correlation matrix composed of two factors, a double factor and a unique variable.}
\format{
A 10 by 10 correlation matrix
}
\source{
\url{https://github.com/quantmeth}
}
\usage{
ex_3factors_doub_unique
}
\description{
A correlation matrix composed of 10 items based on two main factors among which there is
two cross-loadings. There is also a double factors and an unique variable.
Given by Achim, A. (personal communication).
}
\keyword{datasets}
