% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ex_4factors_corr.R
\docType{data}
\name{ex_4factors_corr}
\alias{ex_4factors_corr}
\title{A correlation matrix composed of 4 correlated factors.}
\format{
A 12 by 12 correlation matrix
}
\source{
\url{https://github.com/quantmeth}
}
\usage{
ex_4factors_corr
}
\description{
A correlation matrix composed of 12 items based on 4 factors with 3 variables
each. Loadings equals to .9, .9, and .3. Factors 1 and 2, and factors 3 and 4 are 
correlated at .7. Given by Achim, A (personal communication).
}
\keyword{datasets}
