% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Ledermann.R
\name{Ledermann}
\alias{Ledermann}
\title{Ledermann bound.}
\usage{
Ledermann(p)
}
\arguments{
\item{p}{The number of variable.}
}
\value{
The Ledermann bound.
}
\description{
Returns the maximum number of latent factors in a factor analysis model.
}
\examples{
Ledermann(ncol(ex_4factors_corr))
}
\references{
Ledermann, W. (1937). On the rank of reduced correlation matrices in multiple factor analysis. \emph{Psychometrika}, {2}, 85–93.
}
\author{
André Achim (Matlab)
P.-O. Caron (R)
}
