% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MAP.R
\name{MAP}
\alias{MAP}
\alias{map}
\title{Minimum average partial correlation (MAP)}
\usage{
MAP(.data, ...)
}
\arguments{
\item{.data}{a data frame, a numeric matrix, covariance matrix or correlation matrix from which to determine the number of factors.}

\item{...}{further argument for \code{cor_nest()}.}
}
\value{
The number of factors to retain.
}
\description{
Minimum average partial correlation (MAP)
}
\examples{
D <- genr8(n = 42, R = ex_4factors_corr)
MAP(D)
}
\references{
Velicer, W. F. (1976). Determining the number of components from the matrix of partial correlations. \emph{Psychometrika}, \emph{41}(3), 321-327. \doi{10.1007/BF02293557}
}
