% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meek_bouchard.R
\docType{data}
\name{meek_bouchard}
\alias{meek_bouchard}
\title{A correlation matrix given by Meek-Bouchard.}
\format{
A 44 by 44 correlation matrix
}
\source{
\url{https://github.com/quantmeth}
}
\usage{
meek_bouchard
}
\description{
An empirical correlation matrix composed of 44 items given by Meek-Bouchard, C. (personal communication).
}
\keyword{datasets}
