\name{nestingparameters}
\alias{nestingparameters}
\docType{data}
\title{
Canadian Bird Nesting Parameters}
\description{
Nesting parameters of birds nesting in Canada}
\usage{data(nestingparameters)}
\format{
  A data frame with 2100 observations on the following 12 variables.
  \describe{
    \item{\code{common_name}}{a character vector}
    \item{\code{clutch}}{a numeric vector}
    \item{\code{cmin}}{a numeric vector}
    \item{\code{cmax}}{a numeric vector}
    \item{\code{laying}}{a numeric vector}
    \item{\code{incub}}{a numeric vector}
    \item{\code{start}}{a numeric vector}
    \item{\code{rearing}}{a numeric vector}
    \item{\code{sp}}{a character vector}
    \item{\code{scientific_name}}{a character vector}
    \item{\code{nom_commun}}{a character vector}
    \item{\code{species_code}}{a character vector}
  }
}
\details{
Values were adapted from the species BNA accounts (Birds of North America).
}
\source{
To be described}
\references{
The Birds of North America Online (A. Poole, Ed.). Ithaca: Cornell Laboratory
of Ornithology; Retrieved from The Birds of North America Online database,
URL: http://bna.birds.cornell.edu/BNA/, AUG 2005.
}

