\name{nestwatch}
\alias{nestwatch}
\docType{data}
\title{
The Nest Watch database}
\description{
Current content of the Nest Watch database}
\usage{data(nestwatch)}
\format{
  A data frame with 513056 observations on the following 42 variables.
  \describe{
    \item{\code{nest_id}}{a numeric vector}
    \item{\code{status_code}}{a character vector}
    \item{\code{visit_id}}{a numeric vector}
    \item{\code{month}}{a numeric vector}
    \item{\code{day}}{a numeric vector}
    \item{\code{time}}{a numeric vector}
    \item{\code{eggs}}{a numeric vector}
    \item{\code{young}}{a numeric vector}
    \item{\code{deadyg}}{a numeric vector}
    \item{\code{bhco_eggs}}{a numeric vector}
    \item{\code{bhco_young}}{a numeric vector}
    \item{\code{bhco_deadyg}}{a numeric vector}
    \item{\code{nest_no}}{a numeric vector}
    \item{\code{user_id}}{a numeric vector}
    \item{\code{year}}{a numeric vector}
    \item{\code{species_code}}{a character vector}
    \item{\code{species_code2}}{a character vector}
    \item{\code{statprov_code}}{a character vector}
    \item{\code{county_code}}{a character vector}
    \item{\code{post_code}}{a character vector}
    \item{\code{square_id}}{a character vector}
    \item{\code{loc_label}}{a character vector}
    \item{\code{loc_type}}{a character vector}
    \item{\code{name}}{a character vector}
    \item{\code{description}}{a character vector}
    \item{\code{coordinates_source}}{a character vector}
    \item{\code{coordinates_type}}{a character vector}
    \item{\code{lat}}{a numeric vector}
    \item{\code{lon}}{a numeric vector}
    \item{\code{utm_zone}}{a numeric vector}
    \item{\code{utm_easting}}{a numeric vector}
    \item{\code{utm_northing}}{a numeric vector}
    \item{\code{altitude}}{a numeric vector}
    \item{\code{data_source}}{a character vector}
    \item{\code{date}}{a character vector}
    \item{\code{type_nw_code}}{a character vector}
    \item{\code{stage_code}}{a character vector}
    \item{\code{outcome}}{a character vector}
    \item{\code{nbvisit}}{a numeric vector}
    \item{\code{initiation}}{a character vector}
    \item{\code{fledging}}{a character vector}
    \item{\code{warning}}{a character vector}
  }
}
\details{
The database presented here has been filtered and modified to be usable by Rnest.}
\source{
To be described}
\references{
BSC. 2011. Project NestWatch. Bird Studies Canada / Études d’oiseaux Canada, URL: http://www.bsc-eoc.org
}

