% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/new.R
\name{new}
\alias{new}
\title{YASR (new - 29-09-2022)}
\usage{
new(
  data,
  n.obs = NULL,
  nreps = 1000,
  alpha = 0.05,
  max.fact = max(which(dof(nv, 0:nv) > (nv/2))) - 1,
  ...
)
}
\arguments{
\item{data}{A data frame, a numeric matrix, covariance matrix or correlation matrix from which to determine the number of factors.}

\item{n.obs}{The number of cases (subjects, participants, or units) if a covariance matrix is supplied in \code{data}.}

\item{nreps}{The number of replications to simulate. Default is 1000.}

\item{alpha}{Type I error rates or \code{(1-alpha)*100\%} confidence intervals. Default is .10.}

\item{max.fact}{An optional maximum number of factor to extract.}

\item{...}{Arguments for \code{method} that can be supplied. See details.}
}
\value{
The number of factors
}
\description{
YASR (new - 29-09-2022)
}
\examples{
\dontrun{
new(ex_4factors_corr, n.obs = 84)
}
}
