% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pa.R
\name{pa}
\alias{pa}
\title{Parallel analysis}
\usage{
pa(
  data = NULL,
  n = nrow(data),
  p = ncol(data),
  nreps = 1000,
  alpha = 0.05,
  ...
)
}
\arguments{
\item{data}{data.frame}

\item{n}{number of subjects}

\item{p}{number of variables}

\item{nreps}{number of replications}

\item{alpha}{type I error rate}

\item{...}{Other arguments}
}
\value{
nfactors (if data is supplied) and sampled eigenvalues
}
\description{
Parallel analysis
}
\examples{
pa(ex_2factors)
E <- pa(n = 10, p = 2, nrep = 5)
}
