\name{predictions}
\alias{predictions}
\docType{data}
\title{
Model predictions
}
\description{
Model predictions
}
\usage{data(predictions)}
\format{
  A data frame with 239074 observations on the following 4 variables.
  \describe{
    \item{\code{ecodistrict}}{a numeric vector}
    \item{\code{type}}{a character vector}
    \item{\code{species_code}}{a character vector}
    \item{\code{jul}}{a numeric vector}
  }
}
\details{
...
}
\source{
...
}
\references{
...
}
\examples{
data(predictions)
...
}
\keyword{datasets}
