% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.nest.R
\name{print.nest}
\alias{print.nest}
\title{Print results of NEST}
\usage{
\method{print}{nest}(x, ...)
}
\arguments{
\item{x}{An object of class "nest".}

\item{...}{Further arguments for other methods, ignored for "nest".}
}
\value{
No return value, called for side effects.
}
\description{
Print the number of factors to retain according to confidence levels.
}
\examples{
results <- nest(ex_2factors, n = 100)
print(results)
}
