\name{rNest-package}
\alias{rNest-package}
\alias{rNest}
\docType{package}
\title{
Estimating Nesting Chronology and Breeding Phenology in Birds
}
\description{
Tools for estimating nesting chronology and describing breeding phenology in birds using nest observations.
}
\details{
\tabular{ll}{
Package: \tab rNest\cr
Type: \tab Package\cr
Version: \tab 0.2\cr
Date: \tab 2014-03-11\cr
License: \tab GPL 2\cr
}
}
\author{
Francois Rousseu, Bruno Drolet <bruno.drolet@ec.gc.ca>

Maintainer: Francois Rousseu <francoisrousseu@hotmail.com>
}
\references{
To be described
}

