% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SEA.R
\name{SEA}
\alias{SEA}
\title{Significant Eigenvalues Assessment}
\usage{
SEA(data, n.obs, alpha = 0.05, nreps = 5000)
}
\arguments{
\item{data}{A data frame, a numeric matrix, covariance matrix or correlation matrix from which to determine the number of factors.}

\item{n.obs}{The number of cases (subjects, participants, or units) if a covariance matrix is supplied in \code{data}.}

\item{alpha}{Type I error rate}

\item{nreps}{The number of replications to simulate. Default is 1000.}
}
\value{
\code{SEA} returns the number of factors to retain.
}
\description{
Significant Eigenvalues Assessment
}
\examples{
SEA(ex_2factors, n.obs = 50)
}
