\name{specodistricts}
\alias{specodistricts}
\docType{data}
\title{
Species ranges
}
\description{
Species ecodistricts association
}
\usage{data(specodistricts)}
\format{
  A data frame with 1021 observations on the following 374 variables.
  \describe{
    \item{\code{ecodis}}{a numeric vector}
    \item{\code{GWFG}}{a logical vector}
    \item{\code{SNGO}}{a logical vector}
    \item{\code{ROGO}}{a logical vector}
    \item{\code{BRAN}}{a logical vector}
    \item{\code{CAGO}}{a logical vector}
    \item{\code{MUSW}}{a logical vector}
    \item{\code{TRUS}}{a logical vector}
    \item{\code{TUSW}}{a logical vector}
    \item{\code{WODU}}{a logical vector}
    \item{\code{GADW}}{a logical vector}
    \item{\code{AMWI}}{a logical vector}
    \item{\code{ABDU}}{a logical vector}
    \item{\code{MALL}}{a logical vector}
    \item{\code{BWTE}}{a logical vector}
    \item{\code{CITE}}{a logical vector}
    \item{\code{NSHO}}{a logical vector}
    \item{\code{NOPI}}{a logical vector}
    \item{\code{GWTE}}{a logical vector}
    \item{\code{CANV}}{a logical vector}
    \item{\code{REDH}}{a logical vector}
    \item{\code{RNDU}}{a logical vector}
    \item{\code{GRSC}}{a logical vector}
    \item{\code{LESC}}{a logical vector}
    \item{\code{KIEI}}{a logical vector}
    \item{\code{COEI}}{a logical vector}
    \item{\code{SUSC}}{a logical vector}
    \item{\code{WWSC}}{a logical vector}
    \item{\code{BLSC}}{a logical vector}
    \item{\code{LTDU}}{a logical vector}
    \item{\code{BUFF}}{a logical vector}
    \item{\code{COGO}}{a logical vector}
    \item{\code{BAGO}}{a logical vector}
    \item{\code{HOME}}{a logical vector}
    \item{\code{COME}}{a logical vector}
    \item{\code{RBME}}{a logical vector}
    \item{\code{RUDU}}{a logical vector}
    \item{\code{CAQU}}{a logical vector}
    \item{\code{NOBO}}{a logical vector}
    \item{\code{RUGR}}{a logical vector}
    \item{\code{GRSG}}{a logical vector}
    \item{\code{SPGR}}{a logical vector}
    \item{\code{WIPT}}{a logical vector}
    \item{\code{ROPT}}{a logical vector}
    \item{\code{WTPT}}{a logical vector}
    \item{\code{DUGR}}{a logical vector}
    \item{\code{STGR}}{a logical vector}
    \item{\code{WITU}}{a logical vector}
    \item{\code{RTLO}}{a logical vector}
    \item{\code{PALO}}{a logical vector}
    \item{\code{COLO}}{a logical vector}
    \item{\code{PBGR}}{a logical vector}
    \item{\code{HOGR}}{a logical vector}
    \item{\code{RNGR}}{a logical vector}
    \item{\code{EAGR}}{a logical vector}
    \item{\code{WEGR}}{a logical vector}
    \item{\code{DCCO}}{a logical vector}
    \item{\code{AWPE}}{a logical vector}
    \item{\code{AMBI}}{a logical vector}
    \item{\code{LEBI}}{a logical vector}
    \item{\code{GBHE}}{a logical vector}
    \item{\code{GREG}}{a logical vector}
    \item{\code{CAEG}}{a logical vector}
    \item{\code{GRHE}}{a logical vector}
    \item{\code{BCNH}}{a logical vector}
    \item{\code{TUVU}}{a logical vector}
    \item{\code{OSPR}}{a logical vector}
    \item{\code{BAEA}}{a logical vector}
    \item{\code{NOHA}}{a logical vector}
    \item{\code{SSHA}}{a logical vector}
    \item{\code{COHA}}{a logical vector}
    \item{\code{NOGO}}{a logical vector}
    \item{\code{RSHA}}{a logical vector}
    \item{\code{BWHA}}{a logical vector}
    \item{\code{SWHA}}{a logical vector}
    \item{\code{RTHA}}{a logical vector}
    \item{\code{FEHA}}{a logical vector}
    \item{\code{RLHA}}{a logical vector}
    \item{\code{GOEA}}{a logical vector}
    \item{\code{AMKE}}{a logical vector}
    \item{\code{MERL}}{a logical vector}
    \item{\code{GYRF}}{a logical vector}
    \item{\code{PEFA}}{a logical vector}
    \item{\code{PRFA}}{a logical vector}
    \item{\code{YERA}}{a logical vector}
    \item{\code{KIRA}}{a logical vector}
    \item{\code{VIRA}}{a logical vector}
    \item{\code{SORA}}{a logical vector}
    \item{\code{COMO}}{a logical vector}
    \item{\code{AMCO}}{a logical vector}
    \item{\code{SACR}}{a logical vector}
    \item{\code{BBPL}}{a logical vector}
    \item{\code{AMGP}}{a logical vector}
    \item{\code{SEPL}}{a logical vector}
    \item{\code{PIPL}}{a logical vector}
    \item{\code{KILL}}{a logical vector}
    \item{\code{BNST}}{a logical vector}
    \item{\code{AMAV}}{a logical vector}
    \item{\code{SPSA}}{a logical vector}
    \item{\code{SOSA}}{a logical vector}
    \item{\code{WATA}}{a logical vector}
    \item{\code{GRYE}}{a logical vector}
    \item{\code{WILL}}{a logical vector}
    \item{\code{LEYE}}{a logical vector}
    \item{\code{UPSA}}{a logical vector}
    \item{\code{WHIM}}{a logical vector}
    \item{\code{LBCU}}{a logical vector}
    \item{\code{HUGO}}{a logical vector}
    \item{\code{MAGO}}{a logical vector}
    \item{\code{RUTU}}{a logical vector}
    \item{\code{SURF}}{a logical vector}
    \item{\code{REKN}}{a logical vector}
    \item{\code{SAND}}{a logical vector}
    \item{\code{SESA}}{a logical vector}
    \item{\code{LESA}}{a logical vector}
    \item{\code{WRSA}}{a logical vector}
    \item{\code{BASA}}{a logical vector}
    \item{\code{PESA}}{a logical vector}
    \item{\code{PUSA}}{a logical vector}
    \item{\code{DUNL}}{a logical vector}
    \item{\code{STSA}}{a logical vector}
    \item{\code{BBSA}}{a logical vector}
    \item{\code{SBDO}}{a logical vector}
    \item{\code{LBDO}}{a logical vector}
    \item{\code{COSN}}{a logical vector}
    \item{\code{AMWO}}{a logical vector}
    \item{\code{WIPH}}{a logical vector}
    \item{\code{RNPH}}{a logical vector}
    \item{\code{REPH}}{a logical vector}
    \item{\code{SAGU}}{a logical vector}
    \item{\code{BOGU}}{a logical vector}
    \item{\code{BHGU}}{a logical vector}
    \item{\code{LIGU}}{a logical vector}
    \item{\code{FRGU}}{a logical vector}
    \item{\code{MEGU}}{a logical vector}
    \item{\code{RBGU}}{a logical vector}
    \item{\code{CAGU}}{a logical vector}
    \item{\code{HERG}}{a logical vector}
    \item{\code{GLGU}}{a logical vector}
    \item{\code{GBBG}}{a logical vector}
    \item{\code{CATE}}{a logical vector}
    \item{\code{BLTE}}{a logical vector}
    \item{\code{COTE}}{a logical vector}
    \item{\code{ARTE}}{a logical vector}
    \item{\code{FOTE}}{a logical vector}
    \item{\code{PAJA}}{a logical vector}
    \item{\code{LTJA}}{a logical vector}
    \item{\code{BLGU}}{a logical vector}
    \item{\code{RODO}}{a logical vector}
    \item{\code{MODO}}{a logical vector}
    \item{\code{YBCU}}{a logical vector}
    \item{\code{BBCU}}{a logical vector}
    \item{\code{FLOW}}{a logical vector}
    \item{\code{EASO}}{a logical vector}
    \item{\code{GHOW}}{a logical vector}
    \item{\code{SNOW}}{a logical vector}
    \item{\code{NHOW}}{a logical vector}
    \item{\code{BUOW}}{a logical vector}
    \item{\code{BDOW}}{a logical vector}
    \item{\code{GGOW}}{a logical vector}
    \item{\code{LEOW}}{a logical vector}
    \item{\code{SEOW}}{a logical vector}
    \item{\code{BOOW}}{a logical vector}
    \item{\code{NSWO}}{a logical vector}
    \item{\code{CONI}}{a logical vector}
    \item{\code{CWWI}}{a logical vector}
    \item{\code{WPWI}}{a logical vector}
    \item{\code{CHSW}}{a logical vector}
    \item{\code{RTHU}}{a logical vector}
    \item{\code{CAHU}}{a logical vector}
    \item{\code{RUHU}}{a logical vector}
    \item{\code{BEKI}}{a logical vector}
    \item{\code{RHWO}}{a logical vector}
    \item{\code{RBWO}}{a logical vector}
    \item{\code{YBSA}}{a logical vector}
    \item{\code{RNSA}}{a logical vector}
    \item{\code{RBSA}}{a logical vector}
    \item{\code{DOWO}}{a logical vector}
    \item{\code{HAWO}}{a logical vector}
    \item{\code{ATTW}}{a logical vector}
    \item{\code{BBWO}}{a logical vector}
    \item{\code{NOFL}}{a logical vector}
    \item{\code{PIWO}}{a logical vector}
    \item{\code{OSFL}}{a logical vector}
    \item{\code{WEWP}}{a logical vector}
    \item{\code{EAWP}}{a logical vector}
    \item{\code{YBFL}}{a logical vector}
    \item{\code{ACFL}}{a logical vector}
    \item{\code{ALFL}}{a logical vector}
    \item{\code{WIFL}}{a logical vector}
    \item{\code{LEFL}}{a logical vector}
    \item{\code{HAFL}}{a logical vector}
    \item{\code{DUFL}}{a logical vector}
    \item{\code{COFL}}{a logical vector}
    \item{\code{EAPH}}{a logical vector}
    \item{\code{SAPH}}{a logical vector}
    \item{\code{GCFL}}{a logical vector}
    \item{\code{WEKI}}{a logical vector}
    \item{\code{EAKI}}{a logical vector}
    \item{\code{LOSH}}{a logical vector}
    \item{\code{NSHR}}{a logical vector}
    \item{\code{WEVI}}{a logical vector}
    \item{\code{YTVI}}{a logical vector}
    \item{\code{BHVI}}{a logical vector}
    \item{\code{HUVI}}{a logical vector}
    \item{\code{WAVI}}{a logical vector}
    \item{\code{PHVI}}{a logical vector}
    \item{\code{REVI}}{a logical vector}
    \item{\code{GRAJ}}{a logical vector}
    \item{\code{STJA}}{a logical vector}
    \item{\code{BLJA}}{a logical vector}
    \item{\code{CLNU}}{a logical vector}
    \item{\code{BBMA}}{a logical vector}
    \item{\code{AMCR}}{a logical vector}
    \item{\code{NOCR}}{a logical vector}
    \item{\code{CORA}}{a logical vector}
    \item{\code{SKLA}}{a logical vector}
    \item{\code{HOLA}}{a logical vector}
    \item{\code{PUMA}}{a logical vector}
    \item{\code{TRES}}{a logical vector}
    \item{\code{VGSW}}{a logical vector}
    \item{\code{NRWS}}{a logical vector}
    \item{\code{BANS}}{a logical vector}
    \item{\code{CLSW}}{a logical vector}
    \item{\code{BARS}}{a logical vector}
    \item{\code{BCCH}}{a logical vector}
    \item{\code{MOCH}}{a logical vector}
    \item{\code{CBCH}}{a logical vector}
    \item{\code{BOCH}}{a logical vector}
    \item{\code{TUTI}}{a logical vector}
    \item{\code{BUSH}}{a logical vector}
    \item{\code{RBNU}}{a logical vector}
    \item{\code{WBNU}}{a logical vector}
    \item{\code{PYNU}}{a logical vector}
    \item{\code{BRCR}}{a logical vector}
    \item{\code{ROWR}}{a logical vector}
    \item{\code{CANW}}{a logical vector}
    \item{\code{CARW}}{a logical vector}
    \item{\code{BEWR}}{a logical vector}
    \item{\code{HOWR}}{a logical vector}
    \item{\code{WIWR}}{a logical vector}
    \item{\code{SEWR}}{a logical vector}
    \item{\code{MAWR}}{a logical vector}
    \item{\code{BGGN}}{a logical vector}
    \item{\code{AMDI}}{a logical vector}
    \item{\code{GCKI}}{a logical vector}
    \item{\code{RCKI}}{a logical vector}
    \item{\code{NOWH}}{a logical vector}
    \item{\code{EABL}}{a logical vector}
    \item{\code{WEBL}}{a logical vector}
    \item{\code{MOBL}}{a logical vector}
    \item{\code{TOSO}}{a logical vector}
    \item{\code{VEER}}{a logical vector}
    \item{\code{GCTH}}{a logical vector}
    \item{\code{BITH}}{a logical vector}
    \item{\code{SWTH}}{a logical vector}
    \item{\code{HETH}}{a logical vector}
    \item{\code{WOTH}}{a logical vector}
    \item{\code{AMRO}}{a logical vector}
    \item{\code{VATH}}{a logical vector}
    \item{\code{GRCA}}{a logical vector}
    \item{\code{NOMO}}{a logical vector}
    \item{\code{BRTH}}{a logical vector}
    \item{\code{EUST}}{a logical vector}
    \item{\code{AMPI}}{a logical vector}
    \item{\code{SPPI}}{a logical vector}
    \item{\code{BOWA}}{a logical vector}
    \item{\code{CEDW}}{a logical vector}
    \item{\code{LALO}}{a logical vector}
    \item{\code{CCLO}}{a logical vector}
    \item{\code{SMLO}}{a logical vector}
    \item{\code{MCLO}}{a logical vector}
    \item{\code{SNBU}}{a logical vector}
    \item{\code{OVEN}}{a logical vector}
    \item{\code{LOWA}}{a logical vector}
    \item{\code{NOWA}}{a logical vector}
    \item{\code{GWWA}}{a logical vector}
    \item{\code{BWWA}}{a logical vector}
    \item{\code{BAWW}}{a logical vector}
    \item{\code{PROW}}{a logical vector}
    \item{\code{TEWA}}{a logical vector}
    \item{\code{OCWA}}{a logical vector}
    \item{\code{NAWA}}{a logical vector}
    \item{\code{CONW}}{a logical vector}
    \item{\code{MGWA}}{a logical vector}
    \item{\code{MOWA}}{a logical vector}
    \item{\code{COYE}}{a logical vector}
    \item{\code{HOWA}}{a logical vector}
    \item{\code{AMRE}}{a logical vector}
    \item{\code{CMWA}}{a logical vector}
    \item{\code{CERW}}{a logical vector}
    \item{\code{NOPA}}{a logical vector}
    \item{\code{MAWA}}{a logical vector}
    \item{\code{BBWA}}{a logical vector}
    \item{\code{BLBW}}{a logical vector}
    \item{\code{YWAR}}{a logical vector}
    \item{\code{CSWA}}{a logical vector}
    \item{\code{BLPW}}{a logical vector}
    \item{\code{BTBW}}{a logical vector}
    \item{\code{PAWA}}{a logical vector}
    \item{\code{PIWA}}{a logical vector}
    \item{\code{YRWA}}{a logical vector}
    \item{\code{PRAW}}{a logical vector}
    \item{\code{BTYW}}{a logical vector}
    \item{\code{TOWA}}{a logical vector}
    \item{\code{BTNW}}{a logical vector}
    \item{\code{CAWA}}{a logical vector}
    \item{\code{WIWA}}{a logical vector}
    \item{\code{YBCH}}{a logical vector}
    \item{\code{SPTO}}{a logical vector}
    \item{\code{EATO}}{a logical vector}
    \item{\code{ATSP}}{a logical vector}
    \item{\code{CHSP}}{a logical vector}
    \item{\code{CCSP}}{a logical vector}
    \item{\code{BRSP}}{a logical vector}
    \item{\code{FISP}}{a logical vector}
    \item{\code{VESP}}{a logical vector}
    \item{\code{LASP}}{a logical vector}
    \item{\code{LARB}}{a logical vector}
    \item{\code{SAVS}}{a logical vector}
    \item{\code{GRSP}}{a logical vector}
    \item{\code{BAIS}}{a logical vector}
    \item{\code{HESP}}{a logical vector}
    \item{\code{LCSP}}{a logical vector}
    \item{\code{NSTS}}{a logical vector}
    \item{\code{FOSP}}{a logical vector}
    \item{\code{SOSP}}{a logical vector}
    \item{\code{LISP}}{a logical vector}
    \item{\code{SWSP}}{a logical vector}
    \item{\code{WTSP}}{a logical vector}
    \item{\code{HASP}}{a logical vector}
    \item{\code{WCSP}}{a logical vector}
    \item{\code{GCSP}}{a logical vector}
    \item{\code{DEJU}}{a logical vector}
    \item{\code{SCTA}}{a logical vector}
    \item{\code{WETA}}{a logical vector}
    \item{\code{NOCA}}{a logical vector}
    \item{\code{RBGR}}{a logical vector}
    \item{\code{BHGR}}{a logical vector}
    \item{\code{LAZB}}{a logical vector}
    \item{\code{INBU}}{a logical vector}
    \item{\code{DICK}}{a logical vector}
    \item{\code{BOBO}}{a logical vector}
    \item{\code{RWBL}}{a logical vector}
    \item{\code{EAME}}{a logical vector}
    \item{\code{WEME}}{a logical vector}
    \item{\code{YHBL}}{a logical vector}
    \item{\code{RUBL}}{a logical vector}
    \item{\code{BRBL}}{a logical vector}
    \item{\code{COGR}}{a logical vector}
    \item{\code{BHCO}}{a logical vector}
    \item{\code{OROR}}{a logical vector}
    \item{\code{BUOR}}{a logical vector}
    \item{\code{BAOR}}{a logical vector}
    \item{\code{GCRF}}{a logical vector}
    \item{\code{PIGR}}{a logical vector}
    \item{\code{PUFI}}{a logical vector}
    \item{\code{CAFI}}{a logical vector}
    \item{\code{HOFI}}{a logical vector}
    \item{\code{RECR}}{a logical vector}
    \item{\code{WWCR}}{a logical vector}
    \item{\code{CORE}}{a logical vector}
    \item{\code{HORE}}{a logical vector}
    \item{\code{PISI}}{a logical vector}
    \item{\code{AMGO}}{a logical vector}
    \item{\code{EVGR}}{a logical vector}
    \item{\code{HOSP}}{a logical vector}
    \item{\code{HADU}}{a logical vector}
    \item{\code{GRAP}}{a logical vector}
    \item{\code{RNEP}}{a logical vector}
    \item{\code{LESP}}{a logical vector}
    \item{\code{BANO}}{a logical vector}
    \item{\code{GFLY}}{a logical vector}
    \item{\code{EYWA}}{a logical vector}
  }
}
\details{
...
}
\source{
...
}
\references{
...
}
\examples{
data(specodistricts)
...
}
\keyword{datasets}
