% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/steiger.R
\name{steiger}
\alias{steiger}
\title{Yet Another Stopping Rule - Steiger}
\usage{
steiger(
  data,
  n.obs = NULL,
  alpha = 0.05,
  max.fact = max(which(dof(ncol(data)) > 0)) - 2,
  ...
)
}
\arguments{
\item{data}{A data frame, a numeric matrix, covariance matrix or correlation matrix from which to determine the number of factors.}

\item{n.obs}{The number of cases (subjects, participants, or units) if a covariance matrix is supplied in \code{data}.}

\item{alpha}{Type I error rate.}

\item{max.fact}{An optional maximum number of factor to extract.}

\item{...}{Arguments for methods.}
}
\value{
\code{YASR} returns the number of factors to retain.
}
\description{
Yet Another Stopping Rule - Steiger
}
\examples{
YASR(ex_2factors, n.obs = 50)
}
