% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.nest.R
\name{summary.nest}
\alias{summary.nest}
\title{Summary results of NEST}
\usage{
\method{summary}{nest}(object, ...)
}
\arguments{
\item{object}{An object of class "nest".}

\item{...}{Further arguments for other methods, ignored for "nest".}
}
\value{
No returned value, called for side effects.
}
\description{
summary method for class "nest".
}
\examples{
results <- nest(ex_2factors, n = 100)
summary(results)
}
