% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tabachnick_fidell2019.R
\docType{data}
\name{tabachnick_fidell2019}
\alias{tabachnick_fidell2019}
\title{A covariance matrix composed of 11 variables.}
\format{
A 11 by 11 covariance matrix.
}
\source{
\url{https://github.com/quantmeth}
}
\usage{
tabachnick_fidell2019
}
\description{
A sample covariance matrix composed of 11 items based on two factors according to Tabachnick and Fidell (2019, see, 576-578).
The first five variables are related to "Verbak IQ", the next five are related to "Performance IQ".
The last variable \code{CODING} is unique. Loadings range between .39 and .76.
}
\references{
Tabachnick, B. G., & Fidell, L. S. (2019). \emph{Using multivariate statistics}. Allyn and Bacon. p. 576-577.
}
\keyword{datasets}
