% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TEST1.R
\docType{data}
\name{TEST1}
\alias{TEST1}
\title{A problematic correlation matrix from Paquin.}
\format{
A correlation matrix in a data.frame.
}
\source{
\url{https://github.com/quantmeth}
}
\usage{
TEST1
}
\description{
A sample correlation matrix composed of 72 items with poorly defined factorial structure. NEST does not converge to a solution and reach very large number of factors.
The associated sample size is 348.
To investigate.
}
\references{
Paquin, S. (personal communication).
}
\keyword{datasets}
