% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unique_variable.R
\name{unique_variable}
\alias{unique_variable}
\title{Probability of unique variables}
\usage{
unique_variable(.data, n = NULL, ...)
}
\arguments{
\item{.data}{a data frame, a numeric matrix, covariance matrix or correlation matrix from which to determine the number of factors.}

\item{n}{the number of cases (subjects, participants, or units) if a covariance matrix is supplied in \code{.data}.}

\item{...}{further arguments for \code{cov_nest()}.}
}
\value{
A data frame containing the F-values and probabilities of the variable to be an unique variable.
}
\description{
Probability of unique variables
}
\examples{
exData <- genr8(n = 420, R = ex_3factors_doub_unique)
unique_variable(exData)
}
\author{
P.-O. Caron (R)
André Achim (Matlab)
}
