% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/YASR1.58.R
\name{YASRnew}
\alias{YASRnew}
\title{Yet Another Stopping Rule (1.58)}
\usage{
YASRnew(
  data,
  n.obs = NULL,
  alpha = 0.05,
  nreps = 1000,
  max.fact = max(which(dof(nv) > (nv/2))) - 1,
  ...
)
}
\arguments{
\item{data}{A data frame, a numeric matrix, covariance matrix or correlation matrix from which to determine the number of factors.}

\item{n.obs}{The number of cases (subjects, participants, or units) if a covariance matrix is supplied in \code{data}.}

\item{alpha}{Type I error rate.}

\item{nreps}{Number of resamples.}

\item{max.fact}{An optional maximum number of factor to extract.}

\item{...}{Arguments for methods.}
}
\value{
\code{YASR} returns the number of factors to retain.
}
\description{
Yet Another Stopping Rule (1.58)
}
\examples{
\dontrun{
YASR1.58(gen_data(x_4factors_corr))
}
}
