% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/achim.R
\docType{data}
\name{achim}
\alias{achim}
\title{A list of seven correlation matrices.}
\format{
A 12 by 12 correlation matrix
}
\source{
\url{https://github.com/quantmeth}
}
\usage{
achim
}
\description{
A a list of seven correlation matrices. Given by Achim, A (personal communication).
}
\keyword{datasets}
