% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/briggs_maccallum2003.R
\docType{data}
\name{briggs_maccallum2003}
\alias{briggs_maccallum2003}
\title{A list of three correlation matrice.}
\format{
A a list of three correlation matrices found in Briggs & MacCullum (2003).
}
\source{
\url{https://github.com/quantmeth}
}
\usage{
briggs_maccallum2003
}
\description{
See Briggs, N. E., & MacCallum, R. C. (2003). Recovery of weak common factors by M,ximum likelihood and ordinary least squares estimation. \emph{Multivariate Nehavioral Research}, \emph{38}(1), 25–56. \url{https://doi.org/10.1207/S15327906MBR3801_2}
}
\keyword{datasets}
