% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cormat.l.R
\docType{data}
\name{cormat.l}
\alias{cormat.l}
\title{A list containing 120 lists of correlation matrices and their underlying characteristics}
\format{
A list containing 120 matrices
}
\source{
\url{https://github.com/quantmeth}
}
\usage{
cormat.l
}
\description{
A list containing 120 lists of \eqn{24 \times 24} correlation matrices 
(R) built to represent different factor structures.
Different levels of loadings (delta, .4, .5, .6, .7, .8), 
correlation between factors (corrfact, .0, .1, .2 .3), and.
number of factors (nfactors, 1:8) are used. The list contained 
matrice (R), and their underlying characteristics (delta, corrfact, 
and nfactors).
}
\keyword{datasets}
