% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cormat.R
\docType{data}
\name{cormat}
\alias{cormat}
\title{A list containing 120 correlation matrices}
\format{
A a list of 120 correlation matrices
}
\source{
\url{https://github.com/quantmeth}
}
\usage{
cormat
}
\description{
A list containing 120 \eqn{24 \times 24} correlation matrices 
(R) built to represent different factor structures.
Details are found in the `cormat.l` data.
}
\keyword{datasets}
