% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pa.R
\name{pa}
\alias{pa}
\title{Parallel analysis}
\usage{
pa(
  data = NULL,
  n = NULL,
  p = NULL,
  nrep = 1000,
  alpha = 0.05,
  crit = NULL,
  ...
)
}
\arguments{
\item{data}{data.frame.}

\item{n}{number of subjects.}

\item{p}{number of variables.}

\item{nrep}{number of replications.}

\item{alpha}{type I error rate.}

\item{crit}{Critical values to compare the eigenvalues.}

\item{...}{Other arguments}
}
\value{
nfactors (if data is supplied) and sampled eigenvalues
}
\description{
Parallel analysis
}
\examples{
pa(ex_2factors, n = 42)
E <- pa(n = 10, p = 2, nrep = 5)
}
